ai_video_pdf_dashboard/requirements.txt
flask
python-dotenv
requests
PyPDF2
faster-whisper
transformers
torch
sentencepiece
